Bug Fix #2234 Custom Form - default value on the select question
The default value of selection type form question should be preserved although the order of options change.

Previously, the default value, which is an serialized array, simply contains the index(es) of option selections posted regardless of the order changes. Consequently, if the user changes the order of the options, then default value also changes accordingly.

Now, it checks the value of the selection instead of only index numbers so that it can maintain the correct selections.

Plus a bit of snoopy fixes for 3.10