Significant changes to the LDAP Bridge.
In some LDAP systems, the user defines which groups it is a part of, and this is how LDAP Bridge worked when getting parents of the asset. However, in other systems, the group defines which users are members, using an attribute in the group. This was previously unhandled, meaning the group cache used by a user did not pick up these LDAP groups. This attribute can now be defined in the Attribute Setup screen.

Also, if this new "Group Members" attribute is provided - either by itself, or in conjunction with the original user-centric attribute - groups containing this attribute can now be expanded. This was previously available under a certain set of circumstances (which included using an Active Directory LDAP and a group attribute of "member"). This default is still used on Active Directory systems if no entry to the new attribute is provided, as AD requires it to be done that way (apparently). The notes on the Attribute Setup page explain what happens when both attributes are provided (this is most applicable to OpenLDAP systems).

Also, the Attribute Setup page has been refactored slightly to split the attributes between user-related and group-related ones.