Bug Fix # 2243 Spell checker and WYSWYG corrupting content
When spell checker plugin starts, it passes the content to spell_parser to check any spelling errors for word by word. During this process it cleans up the posted content with utf8_decode() function and the parser return the value back after utf8_encode() function call. Problem is that the string was never encoded before, thus calling decode function breaks any ISO8859-1 characters (multi octets) within the content.

Even in fact the content does not have to be utf8 encoded since pspell functions are not initialized with any specific encoding scheme.

This bug was never been discovered until the changes in WYSIWYG was made to keep multi octet characters (i.e. unicode chars) as it is in the database to enable the search work on non-english languages.

The same bug fix applied to all 3 versions.