Bug Fix #2128 WYSIWYG Editor can't handle &deg; from "Special Characters"
As mentioned in the bug report, this bug is oracle specific. When HTML Tidy converts the given HTML according to the encoding scheme such as 'utf8' or 'latin1', it converts HTML entity to a raw encoded string.

For example, if the user types '&deg;', then Tidy converts it as ASCII value 194 + ASCII value 176, which is UTF8 encoded degree sign. In Postgres this multi octets string can be stored and retrieved without any corruption, however, Oracle does not like this type of value and changes ASCII values to 63 ('?').

Further investigation is required with regards to Oracle DB encoding scheme, however, this fix will give an option not to pass character encoding option to HTML Tidy.

The admin user can set 'Enable Character Encoding' to 'no' in External Tools Configuration, i.e. '&deg;' will be stored as '&deg;'.

The default value is 'yes'.

This fix will only be committed to the current development version of matrix.