Bug Fix 2376
The customised Asset Map Root Asset was generating error message in the asset map or the map is expanded but failed to teleport.

The cause was that the rootlineage parameter passed to Asset Map was incorrectly generated. Previously, it simply picked the first lineage among the ones returned from asset manager. The problem was that getLinkLineages() function in asset manager returns all lineages including Notice links and rootlineage string could be formed with notice link, which does not work.

Now, it picks the first lineage among SQ_SC_LINK_BACKEND_NAV link types.

Also, note that when the rootlineage is specified, the initial lineage selection does  not work. For example, you can add '/_admin' at the end of any asset URL, then the asset map automatically expand to the asset. This is  the feature hasn't been implemented in asset map.