Bug Fix #2388 Apostrophe Searching does not work
When a word with special characters including apostrophe is indexed, the search manager remove them before the word is indexed. However, when the search is performed, those special characters were passed, which will prevent from getting proper result.

Now, when the search query is typed from the user, the query text will be cleaned up as the same way Matrix indexes. For example, if the user's surname is "O'brian", then "obrian" will be indexed and searched too.