Performance Improvement for Related Asset Listing
Re-factoring getAssetList() function in Page Asset Listing in order to split the functionalities of getting the list of assets and sorting them.

The performance problem found in Related Asset Listing with Metadata sorting option is on. Since it gets the list of assets from the parent getAssetList() function call which includes the sorting, it has to wait extra processing time for sorting unnecessary assets before it filters by tag and this waiting time can be significant depending on the number of assets listed.

Now, _getAssetList() and _getSortedAssetList() added so that it can  perform the job separately, furthermore Related Asset Listing can filter the list by tags before it goes into sorting.