Oracle empty string fix in createAssetLink() function.
The query was performing string comparison with the literal '', which is treated as null in Oracle.

To demonstrate this, just try the following two queries in Oracle.

SELECT 'Hello World' FROM DUAL WHERE 'aaa' > 'a';
SELECT 'Hello World' FROM DUAL WHERE 'aaa' > '';

and in Postgres

SELECT 'Hello World' WHERE 'aaa' > 'a';
SELECT 'Hello World' WHERE 'aaa' > '';

The fix is to omit the where condition if the treeid is an empty string.