Bug Fix #2494 Link reports generates warnings on valid named achor tags.
The A element can have a link (with href) and create a named anchor (with name) simultaneously, which is XHTML 1.0 Strict valid. e.g)

<a id="hello" name="hello" href="http://www.hello.com" />

However, the broken link report can not match this particular format as a valid named anchor tag, resulting in reporting error message saying,
'Link points to valid asset, but the anchor "hello" does not exist'.

The fix is to  modify the regular expression which matches the named anchor tag to allow 'A' element to have 'href' attribute.