Major internal changes to File Upload form question, to handle file uploads made in the middle of incomplete submissions on a multiple-page form.
Files are now stored in a temporary location, which is based partly on the incomplete submission ID for a logged-in user, or the session ID for an anonymous submission. A default filesystem path is provided, but the parent can provide a function to set the web path if it so desires. This web path must be as unique as possible (based on submission and question ID) as it expects to be a directory with only one file.

When the submission is completed, it will move the file to either a File asset (as was the original behaviour), or where no create locations exist, it will leave it where it is. (This is new - previously, a Create Location was required.) In the latter case, it is up to the parent asset (ie. the form) to move it to a more permanent location, and currently, to mark its location inside the Extra Data - files moved to a permanent location will be downloadable from the submission's Details screen.

Submission rule checking, including overwrite rule checking, is now moved to the on-submit portion; this is to avoid situations where a file is valid when initially uploaded, but fails an overwrite rule when finally submitted, or vice versa.

This question type now stores "extra data" to help determine where its uploaded file actually is; it's used in the summary. This was necessary to break the reliance on Create Locations - which would break when a file got moved from there, or if the question's Create Location was changed (now the asset ID is stored, when placed in a create location, or the filesystem path inside the data directory). Where Extra Data does not exist (for older submissions), the previous method of checking underneath the current create location is used.

Explanation for the Overwrite Rule has been expanded. This needed to be updated because it no longer applies when a Create Location is not specified. Overwrite Rule handling has now also been tightened so that it only overwrites files (ie. it will not try to overwrite an asset with the same web-path that is not a File), and it now does its checking for clashes based on web paths, not asset names (previously, it could clash with another web path that it couldn't find because it was only looking for Files). Note that no conflict resolution is used when two people with incomplete submissions upload the same file name within a short space of time - take note when selecting your overwrite rule (Rename is recommended).