New Cron Job: Manage Incomplete Submissions.
This is a cron job borrowing fairly heavily from the "Manage Pending Accounts" cron job that handles unvalidated users for Account Manager Page. It handles incomplete submissions from a multiple-page custom form, sending reminder emails to their owners and deleting them after a certain time period.

This is a system asset that will run on each Cron run. Each form has two settings: reminder period (or warning period) and expiry period. Once the reminder period times out (since the creation time of the submission), a reminder email will be sent to the creator of the submission, as long as an Email address is set for that user. Once the expiry period times out, the submission will be deleted and lost. The reminder email and time periods are set from a new "Incomplete Submissions" screen in the Custom Form. Either or both can be disabled (which is the default) by setting the relevant duration to zero.

It will store a list of submissions that it has already sent a reminder email to in its data directory, and clean them out whenever they are deleted, or completed. (It can't do what Manage Pending Accounts does, ie. changing links from TYPE_1 to TYPE_2, because the relationship between Incomplete Submissions and submission assets must remain TYPE_3.)