Feature Bazaar: Login as another user feature. Time for a multi-line commit message...
System Admins (and root of course) can now log in as other users for troubleshooting purposes.
A key icon is available next to the "Logout" icon for these users, which brings up a small
textfield for input of a username. Clicking "Login" will then close that user's session and
bring up the secondary account. System Admins cannot log in as root - so it's not worth trying that one!

A modified "Logout" icon is shown once in the secondary account, which when clicked will
log the originating user back in and restore their asset history lineage to minimise
disruption.

The Messaging Service logs all user switching, detailing the name of the originating and
secondary users. These messages are:
	- system.security.login.secondary (upon successful user switching)
	- system.security.login.secondary.failed (upon user switching failure - ie; secondary user not live)
	- system.security.logout.secondary.manual (after logout from the secondary user and login as the originating user)

This feature may only be used for Good(tm), not evil.