Remove the old pgsql_to_oracle & oracle_to_pgsql scripts (didn't work for 3.18) and combine them into one script
The new script handles doing a db conversion either way:
- from postgres to oracle
- from oracle to postgres

The script is run the same way:

php scripts/dev/convert_database.php /path/to/matrix

You need to set a source_dsn and destination_dsn (same format as the db.inc file) in the script and it'll handle the rest (drop sequences, drop indexes, copy data, create indexes, recreate sequences). As with the previous script(s), you need to have run step_02 from the installer on the destination db so it creates everything in the first place.