change the way search queries are put together.
Instead of being a 6-table join, it's broken into 3 sections - a query on the sq_sch_idx table with a subquery that limits which parts of
the asset tree to look at which "intersect"s with the permission checking. lots of code comments to show what's going on & what you end
up with etc as well.