New asset type to list all the tags under a root nodes (and its children). This is primarily designed to create front-end tag coulds from matrix tags.
For performance reasons the features of the listing have been simplified: 1 root node only, no filtering on asset types, no link value filtering, no excluded root nodes and assets, no "up the tree" searches, no custom List Format.

The field "Asset Types to List" is actually a thesaurus term (or descendent). So it is important to understand that there are two levels involved here: the children of the root node and the term it is tagged with. The listing first find all the assets under the root node, and then get all the tags of those assets.

Only the "live" terms applied to children with effective read access for the current user will be listed.

By default the result is sorted by the number of assets a term has been applied to.

A few additional back-end fields have been introduced:
* Filtering the tags by a thesaurus (this is optional, if not defined all tags are relevant)
* First Page Only: this indicates that the listing will not use the pagination. This is an optimisation field only and it should probably be moved up to the listing engine one day as it makes a significant difference when showing a top 10 out of 1000s of assets.
* Maximum Tag Size: Determine the maximum value for the %tag_size% keyword in the Term Type Format. This keyword is typically used in cunjunction with CSS to create tag coulds.

Two new keywords can be used in the Term Type Format bodycopy:
** %tag_count%: this is the number of assets under the root node which have been tagged with the term.
** %tag_size%: this is a relative size, between 1 and [Maximum Tag Size]. Used with CSS.

Note that you can also show the tags of the root node only if you set 0 to min/max depth on the details screen.

Also note that this asset has not been tested with Oracle.

Many things can be improved in the future, like putting back some of the search filtering and restrictions present in the listing engine or devising a more custom calculation for the "size" keyword based on the weight of the tags.