Feature 3522: Form Captcha Enhancements - True Type Fonts and Arc. Code graciously contributed by Nic Hubbard.
Some new options are now available for CAPTCHA fields on the "Details" screen in Asset Builder. These are as follows:

	- An option to specify a TrueType Font (from a File asset) to be used to render the CAPTCHA text
	- Font size for the TTF font
	- CAPTCHA width and height
	- Angle range for each character
	- Minimum and maximum pixel distance between characters
	- Show arc shapes across the text, with a customised arc colour
	- An X ordinate offset for the CAPTCHA

A further enhancement as suggested by Nic Hubbard has also been added which is the ability to
have uppercase letters and numbers as part of the generated CAPTCHA text. This has been implemented
as two further options in the "CAPTCHA" section on the "Form Contents" screen and default to "No".
Certain numbers and uppercase letters are excluded from the random selection as these can
appear ambiguous (eg; zero vs uppercase-O, one vs lowercase-L).

If enabled, at least one of the specified character types will appear in the string presuming the
CAPTCHA is at least 3 characters long - and it would want to be   :-)