Bug Re-fix 3483 (resolution for issue raised as Bug 3552: Second-level grouping on a Metadata Field or Standard Field in Asset Listing causes listing to be blank). Code introduced in Bug Fix 3483 was recalculating the 'Immediate Child Group Count' causing issues when displaying second-level groupings.
Furthermore, an earlier change to use dal_array_flatten() instead of array_flatten() when determining assets to display and count was causing the 'Total Child Count' to be returned as zero when accessed from a Search Page (Asset Listing was fine). At the time I cannot see a reason for the change and as it was not a bug when committed has been reverted to restore the expected functionality. Unit Test modifications were made to Asset Listing and Search Page grouping to expect the same totals for the Immediate and Total counts as replicated on a 3.16.8 system.

I feel I have provided sufficient testing to confidently revert the changes made. The 3.18 and 3.20 resolutions have been committed after verifying the results of Listing Engine-based Unit Tests this morning.