Commit Phayl. Modifications to Account Manager keyword dropdown go with recent CAPTCHA modifications, original commit message below...
Added Accessible CAPTCHA Functionality. Also brought recent CAPTCHA modifications (colours, fonts, numbers etc.) into Account Manager.

A new switch is provided in the "CAPTCHA" section of  the "Details" screen of Account Manager and Asset Builder assets and the
"Contents" screen of Custom Form to enable an Accessible CAPTCHA. When enabled, a link will appear next to the CAPTCHA image, whether used
in isolation or within a "%form_captcha%", to toggle between using the CAPTCHA image or an email address field.

When clicked, the image CAPTCHA interface will be replaced with a text field and button, prompting for input of the user's email address.
A verification email will be sent to the specified address once the "Submit Address" button is clicked. The user's session and the
resultant email is populated with a key which is used to match up the intention to submit the form with the corresponding user.

A link is provided in the email which when clicked will verify the key against that stored in the session and flag the form as verified
for CAPTCHA purposes. Once verified, the same key cannot be used again (unless you are extremely lucky - notify us, then buy a lottery ticket).

Once the form is submitted, the verification flag is cleared, requiring the user to start the Accessible CAPTCHA process again for subsequent forms
or upon verification failure of fields in the same form - as is the case with the traditional CAPTCHA.