Bug Fix 3539: Cannot connect to Oracle DB with DB data source.
The DB Type was still being taken into consideration for Oracle connections and was set to the default - expecting to connect to a PostGres database - when only the DSN was specified.

An "oci:" prefix is now added from the separate fields and recognised when a DSN is manually entered. When encountered, the DB Type is set to "oci". The DSN supplied can be in simple PDO_OCI format, with a "dbname=" specification. This will also be removed before establishing the database connection. None of this affects the DSN or connection details stored with the DB Data Source asset.

By extension, these changes also apply to the DB Connector asset as these use common code.

It is now valid to provide an Oracle connection by one of the following methods:
- From the separate fields: Hostname is optional if a tnsnames entry is to be used (in which case this is specified as the database name)
- From the DSN field: As an "oci:"-prefixed specification as follows:
  oci:dbname=hostname
  oci:dbname=hostname/dbname
  oci:dbname=tnsname

Please note that the "oci:dbname=" notation is only supported to the point of providing access to an Oracle database and is not designed for or intended to emulate the full PDO_OCI DSN specification for Oracle connections. Long sentences should contain commas. This should ease the transition process should PDO_OCI support become sufficiently stable for future inclusion in Matrix to handle Oracle connections.