Rework the generateGetParentsQuery method. Instead of doing a "LIKE" search across the whole ast_lnk_tree table, it does things differently depending on the situation (whether min_/max_height vars are available etc).
If possible it will do an exact match on the treeid rather than a like and also remove the length check (since it's doing an exact match, the length check is pointless). This should make most cases perform better but there will still be situations where it can't be done (min_/max_height's are different for example).

plenty of inline comments abound about what it's trying to do.