Calendar Page date range limiting while browsing. A lower and upper bound added to the "Details" screen of the Calendar Page asset to restrict navigation between time period views using the "Next" and "Previous" links.
This restriction is disabled by default, but when enabled will it will be evaluated based on the current system date. Manually entered dates in the address bar outside the relevant range for the view will show no "Next" and "Previous" navigation.

Each view can be configured independently and asymmetrically. For example, the Month view can have no restriction while the Year view is restricted to between 2 years in the past to 4 years in the future.