In order to process and to securely send data to web services we have had too make a significant patch to Matrix to add modifiers to keywords (as supported on Mini), These allow constructions such as  %asset_url~urlencode% for example. These turned out to be essential for sending items such as URLs to web service APIs where they must be encoded for them to work, and in order to safely use returned values to prevent injection accounts and so on. Without this there are large security and functionality issues. We also added some basic modifiers such as filtering x words that are much in demand for general implementation as well as the specific encoding and decoding that are needed for typical API calls. In addition we found it necessary to add xpath basic support in order to pull out single results from XML methods. Modifiers can be chained so multiple ones can be used at once.
This patch is carefully constructed to be low impact on core code and not to affect any code paths where modifiers are not used. The way keywords are used is not changed but a second pass is used to process any modifiers after the normal replacements. This has been tested in a wide variety of keyword usage situations.

Although this is a significant feature enhancement we have been very careful to make it low impact on the code, and it is absolutely necessary to securely and effectively use web services. It is easy to add more functions to the supported list, within the syntactic constraints of keywords.