Large change to the base Asset class' storage of variables, to accommodate the storage of multiple contexts' worth of attribute values. The values of the $vars array now contains an array of values rather than one: one key 'default' contains the base default value defined in the management classes, while numeric keys define the values in various contexts ("0" is the default context).
If a value has not been set in an alternate context, the attribute getter functions will use the default context first, before using the base default. The getter/setter functions now accept a context ID to use (which defaults to the currently active context for the session - or the default session for CLI). Also, if an attribute is marked as "not contextable", the relevant functions will always use the default context.

Also changed the attrByRef() method so that the NULL return there is not returned as a scalar, which will throw a PHP complaint.

Requires some changes in assets that create shadow assets, such as form or bridges.