Reverted original attempt of working with contextual attributes, mainly due to performance concerns, and a desire for greater transparency of the context layer for those working down the line.
The new system stores ALL contexts' values in the sq_ast_attr_val table, rather than only those set, with a new 'use default' column allowing us to quickly cascade changes to the default context to all alternate contexts. This trade-off allows the asset to be loaded faster, as the query for loading the vars array does not change significantly (and is a lot better than the first attempt).

Since contexts should not change outside of the selection in mysource.inc (or in deleting of contexts, but this change happens when no assets are being edited), this change should not break integrity of assets (by loading them in one context and saving them in another). It also simplifies handling of asset vars, as they as now are back to the previous format (ie. no intervening contextid to worry about), apart from the 'is_contextable' setting which remains.

It is planned that access to attributes in another context will still be available, but from Asset Manager, not from Asset itself. This is important because some decisions (such as name and short_name being based on default context values always) are not changing. A fourth param has been added to Asset_Manager::getAssetAttributesByName() to handle this.

New DAL queries are provided for inserting or deleting attribute/unique attribute data when creating or deleting attributes. Attributes will be copied from the default context (and marked as 'using default' it until changed).

Changes require a new field (use_default = char(1) default '1') to be added to the sq_ast_attr_val and sq_ast_attr_uniq_val tables (and rollback versions thereof where appropriate). If contexts have been created in systems before this, they will not receive the copied values unless they are deleted and re-created, and thus the inheriting values may not work properly (ie. may get blank values instead).