Moved the main saveAttributes() queries to DAL, to help Oracle compatibility. However, to do this, some changes needed to be made to the baker for MatrixDAL.
The problem related to an INSERT..SELECT setup where Matrix was unable to send a parameter in the required CLOB format. Previously, this would only be done if the baker determined that a bind variable was being bound to a CLOB column - which DAL is currently unable to do for INSERT..SELECT.

The changes allow temporary OCI-Lob objects - usually created using oci_new_descriptor() - to be sent as a bind variable, where the baker cannot determine whether a bind variable is to be bound to any field in particular. This does not apply to bind variables created by DAL Baker for constant placeholders. The onus is on the calling function to provide a valid OCI-Lob object, including a fresh one for each query (as the LOB object is invalidated when executed). Don't forget that this is only required for Oracle, not for PostgreSQL systems - so if you have to create an OCI-Lob, don't forget to make it conditional on the database being Oracle.

This is intended as a workaround for Oracle and shouldn't be attempted unless absolutely required. If you get errors with bind variables using long (>4000 byte) replacements - resulting in error "can bind a LONG value only for insert into a LONG column" - you can try this.