Add new keywords %card_start_month_select%, %card_start_year_select_XX%, %card_expiry_month_select%, and %card_expiry_year_select_XX% for this Payment Gateway. The new keywords enable selecting months and years from a drop-down box.
If %card_start_year_select% is used without a number, a period of 5 years is used. If a period is specified at the end of the keyword like %card_start_year_select_10%, the specified period (10) is used to render the years in the drop-down box. The same logic is applied for expiry year keyword.