The result of getKeywordReplacement() methods should always be strings. However, PHP is loose typing and we can not force the return value types of getKeywordReplacement() and getXXXKeywordReplacement() methods to string so we should cast the results as string instead.
An example of the failed case is that the getKeywordReplacement() method of an asset return an integer or float of 0. Any string will be evaluated as 0 if we compare it with a number. So the condition ($asset_replacement != "%$matches[1]%") is always evaluated to FALSE if $asset_replacement is 0. Then the keyword replacement is an empty string ('') instead of the number 0.