Split deleteAssetLink() into deleteAssetLinkByLink(). the later function can be called to delete link by passing in the link. deleteAssetLink() can be passed linkid to be deleted ad the parameter, which will in turn call deleteAssetLinkByLink() function.
This is required in case of a ROLLBACK. When a ROLLBACK occurs while creation of an asset, the link are created in the table but the asset is never created. So we have the linkid but if we query getLinkById() with this linkid, it wont return us any links. This ispart of fix to  Fix to #3877 Matrix will create and retain links in DB if asset fails to get created

More comments on function decleration :)