New "multiple approval streams" functionality. A new Workflow Stream asset now exists, which sits between schema and step (a default one is created upon schema creation, and will be added upon asset type install). Those with Administrator permission to an asset will be permitted to apply for approval using the alternative streams.
Most of the workflow schema editing interface has been moved to workflow stream, since this is the most appropriate place for this now. A couple of workflow manager functions now work with a workflow stream instead of schema (since only one stream is installed as a running workflow) - generateWorkflowArray() is the most notable. Others may accept a stream name to run either the default, or a named stream.