slight refactoring of getAsset to make it more readable - return early if we get the asset from the asset_cache.
in getDependantParents and getDependantChildren, cache the path if we can. this will help a little if you are on the "edit contents" screen of a page and it has lots of divs/tables - the path above the page asset (eg the site/folder above it) doesn't change, so we can cache that info.

also in getSnippetKeywords cache the keywords list we create - this will save a little time if you have multiple divs/tables on a page, we don't have to regenerate the list for each asset under the page we're editing, just do it once.