Add "cleanExpiredEntries()" function to the Cache Storage classes.
Necessary for the NZL Feature Bazaar task, "Cron Job to delete expired cache entries." The Cron Job (to be committed next) will call this function, which should clean up any remnants of expired cache entries.

In the case of the Default cache storage (using sq_cache table and filesystem), the function will attempt to delete the cache files from entries that are expired, then delete table entries of those it can successfully delete. Of course, this means the Cron run script needs to be run as the Apache user, since an ordinary user will not have the correct permissions to delete (no group write access is given to the cache bucket directories, which is what is needed to delete).

By default (in base Cache Storage class), no action is taken. In some cases, this is the right thing to do. For instance, Memcache storage has nothing to clean up (at least, nothing we can clean up at a Matrix server level). Onus is on new Cache Storage type creators to override this function, if there is the ability to clean up expired entries.