Feature 4316: Trigger action to update Twitter status.
First commit of this new trigger that allows you to Tweet based on the contents of an asset that fires a trigger.

There is a bit of setting up to do to make this work, due to use of OAuth to authenticate (our only option, since basic auth will be deprecated soon) - the interface tries to explain this. First, you need to have the site registered as an application through Twitter's OAuth developer system to get client (consumer) credentials. After entering these into the action, user needs to go through Twitter's site to associate itself with an account, which must be given permission to post - it will callback to the trigger in Matrix and it will complete the process.

Global keywords and keywords of the triggering asset are supported in the Tweet format. The action supports automatic abbreviation of tweets, in case keyword replacements push the length over 140 characters - links are not abbreviated however, and will throw an error. URL shortening is NOT provided, however if shorter remaps are available in the remap manager (including those provided by "Add Remap" trigger action), it is possible to use a special keyword that will use these if possible.

It may be worth making this trigger non-critical at the moment - as there are a few reasons why tweeting may fail. Length of tweet is one thing (and this is caught by us first), but hitting an update rate limit (1,000 per day, with undisclosed amounts per half-hour) is another. Sending duplicate statuses is also an error for Twitter.

There are some changes made to Trigger Manager in relation to this. The getInterface() function is now passed two additional parameters (params 4 and 5) - a reference to the trigger that the action is a part of, and the action ID (key in the actions attribute for the trigger). This allows you to make modifications to the trigger action in the painting if necessary. This is required in the case of the Tweet trigger action, since it needs to generate and then save request and access tokens in this interface, to reduce the amount of input required from the user to achieve this, and to keep request tokens fresh.