Feature 1404: Apply metadata key and metadata value on different screens. This feature (as re-requested to us) was that the key and value of a select metadata field be made accessible always, regardless of the "visible part" that is exposed to META tags. The keywords will be %asset_metadata_*_key% and %asset_metadata_*_value%, and will be overridden by full field names ending in "_key" and "_value" that result in the same keyword.
Note that Metadata_Field_Select::getValueFromKey() now always returns the value of the key it was sent - it used to return the key straight back (contrary to the function's name) if the "visible part" was the key. This was unsatisfactory for converting the key to value for a "_value" additional keyword, which should not depend on the visible part. Now the check for visible part is made in the functions that use it.