Fix to bug #4475 - WYSIWYG throws Unknow toolbar item 'addkeyword' FIXME error.
The first div wysiwyg content in the Page Content asset is responsible for including the js code for wysiwyg plugins in the page. Replace Keyword is special plugin which is only included if there are relevent keyword replacements for the wysiwyg content asset or its dependent parents/childrens. So if first wysiwyg doesn't requires the Replace Keyword plugin then js code corrsponding to it will not be included in the page, but if any other wysiwyg content in the page requires it then it throw the 'Unknown toolbar item: addkeyword' js alert.

Fixed this by not throwing this error if plugin not found is 'addkeyword'. This doesn't actually fix the issue but only mutes it. Actual fix will involve checking all the div contents in the page for their respective keywords replacement, if any. This will involve significate change in the wysiwyg core regarding how js code for wysiwyg and its plugins are placed on a page. Considering the impact of this bug, which is entirely cosmatic and very unlikelyness of it getting exposed, its not worth fixing.