#4546 - Cannot locate LDAP users in asset map with backslash in CN.Due to the commas in some users' CNs, the ID of the corresponding LDAP user asset in Matrix contains a backslash (to escape the comma and prevent it from causing an LDAP error).
The JavaScript being generated by Matrix in the user's Details screen, however, simply outputs the entire asset ID inside string quotes. This results in a problem since JavaScript interprets the single backslash as an escape sequence and essentially discards it. As a result, the asset locator fails to match the asset ID passed to it by JavaScript with the asset ID in the Matrix.

In case of Info popup, this backslash in JavaScript code causes the js syntax error, causing this popup not to work.