Minor enhancement to UTransaction payment gateway: Previously the products amount in shopping cart had be GST exclusive for UTransact gateway as the GST was calculated and added at gateway's end. However now the UTrasact allows to specify the GST on the transaction.
As per clinet requirement, we set this new field 'GstRate' to zero. This will allow Matrix to calculate and add the tax at the checkout, and send the amount inclusive of GST to the gateway. Since the GST to calculate at gateway end is set to zero, gateway will not add any further GST to the transaction amount.

This means after this change GST rate set on transaction will be based on GST info set on checkout and product. Also 'transaction_gst' keyword replacement will return GST rate set on gateway, and not on the transaction. As we don't add GST at gateway anymore, this keyword servers no more purpose.