Bug #5163: Search log report overwriting the existing entries in Current search report. The search report generator will rotate the search report only if the dates in entries in the current search log file differs more than the given report rollback period. However if report rollback period is set to 24 hours and if the search logs are rotated everyday, then all the entries in the current search log will be within 24 hours time range. Hence, the current search report will never gets rolled-over but simply just overwritten.
For fix, make sure we check last entry in the previously rotated search log, if any, for the comparison with the very first entry in the current
search log, to decide whether to rotate the previous search report file.

NOTE: The current search report generator does not handles (and apparently not designed for) merging of the existing search reports if date(s) in current search log are within the rollback period range of the existing search report. In such case, the current report will still be overwritten by the new data in current search log.