Bug #5294. Updated/New Matrix scripts that does forking and run in batches and/or run multiple threads at once:
system_apply_permissions.php: Takes a permission string in a specific format that allows you to run multiple permissions operations on the specified assets and their children. The script forks and runs in batches to curb memory usage and also runs multiple threads in parallel to speed up the overall operation.

system_apply_metadata_schema.php: Set a metadata schema. Applies a metadata schema to the specified assets and their children (and generates the metadata for that asset). Runs in batches, and multiple threads in parallel.

system_update_lookups.php: Updates lookups for all assets under the specified root node. Forks and runs in batches to keep memory usage down. Also this script request the root node to be a site asset anymore.